+++ test/EventManagerTest.php
@@ -1,26 +1,25 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-eventmanager for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-eventmanager/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-eventmanager/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\EventManager;
+namespace LaminasTest\EventManager;
 
 use ArrayIterator;
-use stdClass;
-use Zend\EventManager\Event;
-use Zend\EventManager\EventInterface;
-use Zend\EventManager\EventManager;
-use Zend\EventManager\SharedEventManager;
-use Zend\EventManager\StaticEventManager;
+use Laminas\EventManager\Event;
+use Laminas\EventManager\EventInterface;
+use Laminas\EventManager\EventManager;
+use Laminas\EventManager\SharedEventManager;
+use Laminas\EventManager\StaticEventManager;
+use Laminas\Stdlib\CallBackHandler;
 use ReflectionClass;
-use Zend\Stdlib\CallBackHandler;
+use stdClass;
 
 /**
- * @group      Zend_EventManager
+ * @group      Laminas_EventManager
  */
 class EventManagerTest extends \PHPUnit_Framework_TestCase
 {
@@ -43,7 +42,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
     public function testAttachShouldReturnCallbackHandler()
     {
         $listener = $this->events->attach('test', [$this, __METHOD__]);
-        $this->assertInstanceOf('Zend\Stdlib\CallbackHandler', $listener);
+        $this->assertInstanceOf('Laminas\Stdlib\CallbackHandler', $listener);
     }
 
     public function testAttachShouldAddListenerToEvent()
@@ -90,7 +89,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
         $this->assertInternalType('array', $listeners);
 
         foreach ($listeners as $listener) {
-            $this->assertInstanceOf('Zend\Stdlib\CallbackHandler', $listener);
+            $this->assertInstanceOf('Laminas\Stdlib\CallbackHandler', $listener);
             $this->assertSame($callback, $listener->getCallback());
         }
     }
@@ -144,7 +143,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
             return str_rot13($string);
         });
         $responses = $this->events->trigger('string.transform', $this, ['string' => ' foo ']);
-        $this->assertInstanceOf('Zend\EventManager\ResponseCollection', $responses);
+        $this->assertInstanceOf('Laminas\EventManager\ResponseCollection', $responses);
         $this->assertEquals(2, $responses->count());
         $this->assertEquals('foo', $responses->first());
         $this->assertEquals(\str_rot13(' foo '), $responses->last());
@@ -168,7 +167,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
             ['string' => 'foo', 'search' => 'f'],
             [$this, 'evaluateStringCallback']
         );
-        $this->assertInstanceOf('Zend\EventManager\ResponseCollection', $responses);
+        $this->assertInstanceOf('Laminas\EventManager\ResponseCollection', $responses);
         $this->assertSame(0, $responses->last());
     }
 
@@ -208,7 +207,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
         $responses = $this->events->trigger('foo.bar', $this, [], function ($result) {
             return ($result === 'found');
         });
-        $this->assertInstanceOf('Zend\EventManager\ResponseCollection', $responses);
+        $this->assertInstanceOf('Laminas\EventManager\ResponseCollection', $responses);
         $this->assertTrue($responses->stopped());
         $result = $responses->last();
         $this->assertEquals('found', $result);
@@ -224,7 +223,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
         $responses = $this->events->trigger('foo.bar', $this, [], function ($result) {
             return ($result === 'found');
         });
-        $this->assertInstanceOf('Zend\EventManager\ResponseCollection', $responses);
+        $this->assertInstanceOf('Laminas\EventManager\ResponseCollection', $responses);
         $this->assertTrue($responses->stopped());
         $this->assertEquals('found', $responses->last());
     }
@@ -238,7 +237,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
         $responses = $this->events->trigger('foo.bar', $this, [], function ($result) {
             return ($result === 'never found');
         });
-        $this->assertInstanceOf('Zend\EventManager\ResponseCollection', $responses);
+        $this->assertInstanceOf('Laminas\EventManager\ResponseCollection', $responses);
         $this->assertFalse($responses->stopped());
         $this->assertEquals('zero', $responses->last());
     }
@@ -267,7 +266,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
     {
         $aggregate = new TestAsset\MockAggregate();
         $method    = $this->events->attachAggregate($aggregate);
-        $this->assertSame('ZendTest\EventManager\TestAsset\MockAggregate::attach', $method);
+        $this->assertSame('LaminasTest\EventManager\TestAsset\MockAggregate::attach', $method);
     }
 
     public function testCanDetachListenerAggregates()
@@ -351,7 +350,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
         $aggregate = new TestAsset\MockAggregate();
         $this->events->attachAggregate($aggregate);
         $method = $this->events->detachAggregate($aggregate);
-        $this->assertSame('ZendTest\EventManager\TestAsset\MockAggregate::detach', $method);
+        $this->assertSame('LaminasTest\EventManager\TestAsset\MockAggregate::detach', $method);
     }
 
     public function testAttachAggregateAcceptsOptionalPriorityValue()
@@ -375,7 +374,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
         $this->events->attach('foo.bar', function ($e) { return 'found'; }, 2);
         $this->events->attach('foo.bar', function ($e) { return 'zero'; }, 1);
         $responses = $this->events->trigger('foo.bar', $this, []);
-        $this->assertInstanceOf('Zend\EventManager\ResponseCollection', $responses);
+        $this->assertInstanceOf('Laminas\EventManager\ResponseCollection', $responses);
         $this->assertTrue($responses->stopped());
         $this->assertEquals('nada', $responses->last());
         $this->assertTrue($responses->contains('bogus'));
@@ -557,21 +556,21 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
     {
         $identifier1 = 'foo';
         $identifiers = [$identifier1];
-        $this->assertInstanceOf('Zend\EventManager\EventManager', $this->events->setIdentifiers($identifier1));
+        $this->assertInstanceOf('Laminas\EventManager\EventManager', $this->events->setIdentifiers($identifier1));
         $this->assertSame($this->events->getIdentifiers(), $identifiers);
         $identifier2 = 'baz';
         $identifiers = [$identifier1, $identifier2];
-        $this->assertInstanceOf('Zend\EventManager\EventManager', $this->events->addIdentifiers($identifier2));
+        $this->assertInstanceOf('Laminas\EventManager\EventManager', $this->events->addIdentifiers($identifier2));
         $this->assertSame($this->events->getIdentifiers(), $identifiers);
     }
 
     public function testIdentifierGetterSettersWorkWithArrays()
     {
         $identifiers = ['foo', 'bar'];
-        $this->assertInstanceOf('Zend\EventManager\EventManager', $this->events->setIdentifiers($identifiers));
+        $this->assertInstanceOf('Laminas\EventManager\EventManager', $this->events->setIdentifiers($identifiers));
         $this->assertSame($this->events->getIdentifiers(), $identifiers);
         $identifiers[] = 'baz';
-        $this->assertInstanceOf('Zend\EventManager\EventManager', $this->events->addIdentifiers($identifiers));
+        $this->assertInstanceOf('Laminas\EventManager\EventManager', $this->events->addIdentifiers($identifiers));
 
         // This is done because the keys doesn't matter, just the values
         $expectedIdentifiers = $this->events->getIdentifiers();
@@ -583,10 +582,10 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
     public function testIdentifierGetterSettersWorkWithTraversables()
     {
         $identifiers = new ArrayIterator(['foo', 'bar']);
-        $this->assertInstanceOf('Zend\EventManager\EventManager', $this->events->setIdentifiers($identifiers));
+        $this->assertInstanceOf('Laminas\EventManager\EventManager', $this->events->setIdentifiers($identifiers));
         $this->assertSame($this->events->getIdentifiers(), (array) $identifiers);
         $identifiers = new ArrayIterator(['foo', 'bar', 'baz']);
-        $this->assertInstanceOf('Zend\EventManager\EventManager', $this->events->addIdentifiers($identifiers));
+        $this->assertInstanceOf('Laminas\EventManager\EventManager', $this->events->addIdentifiers($identifiers));
 
         // This is done because the keys doesn't matter, just the values
         $expectedIdentifiers = $this->events->getIdentifiers();
@@ -629,7 +628,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
                 return true;
             }
         );
-        $this->assertInstanceOf('Zend\Stdlib\CallbackHandler', $callbackHandler);
+        $this->assertInstanceOf('Laminas\Stdlib\CallbackHandler', $callbackHandler);
     }
 
     public function testDoesNotCreateStaticInstanceIfNonePresent()
@@ -698,7 +697,7 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testTriggerThrowsInvalidCallbackExceptionForInvalidCallback()
     {
-        $this->setExpectedException('Zend\EventManager\Exception\InvalidCallbackException',
+        $this->setExpectedException('Laminas\EventManager\Exception\InvalidCallbackException',
         'Invalid callback provided');
 
         $responses = $this->events->trigger(
@@ -711,13 +710,13 @@ class EventManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testAttachThrowsInvalidArgumentExceptionForInvalidCallback()
     {
-        $this->setExpectedException('Zend\EventManager\Exception\InvalidArgumentException');
+        $this->setExpectedException('Laminas\EventManager\Exception\InvalidArgumentException');
         $this->events->attach('*', null);
     }
 
     public function testDetachThrowsInvalidArgumentForInvalidCallback()
     {
-        $this->setExpectedException('Zend\EventManager\Exception\InvalidArgumentException');
+        $this->setExpectedException('Laminas\EventManager\Exception\InvalidArgumentException');
         $this->events->detach(null);
     }