+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#5](https://github.com/zendframework/zend-eventmanager/pull/5) adds a number
+- [zendframework/zend-eventmanager#5](https://github.com/zendframework/zend-eventmanager/pull/5) adds a number
   of unit tests to improve test coverage, and thus maintainability and
   stability.
 
@@ -16,11 +16,11 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- [#3](https://github.com/zendframework/zend-eventmanager/pull/3) removes some
+- [zendframework/zend-eventmanager#3](https://github.com/zendframework/zend-eventmanager/pull/3) removes some
   PHP 5.3- and 5.4-isms (such as marking Traits as requiring 5.4, and closing
   over a copy of `$this`) from the test suite.
 
 ### Fixed
 
-- [#5](https://github.com/zendframework/zend-eventmanager/pull/5) fixes a bug in
+- [zendframework/zend-eventmanager#5](https://github.com/zendframework/zend-eventmanager/pull/5) fixes a bug in
   `FilterIterator` that occurs when attempting to extract from an empty heap.